@@ -1,6 +1,7 @@
 {
   "private": true,
   "sideEffects": false,
+  "type": "module",
   "scripts": {
     "build": "remix build",
     "predev": "rimraf ./public/_redirects",
@@ -9,10 +10,10 @@
     "typecheck": "tsc -b"
   },
   "dependencies": {
-    "@netlify/functions": "^1.3.0",
     "@remix-run/node": "*",
     "@remix-run/react": "*",
     "cross-env": "^7.0.3",
+    "isbot": "^3.7.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
@@ -24,7 +25,7 @@
     "@types/react-dom": "^18.0.8",
     "eslint": "^8.27.0",
     "rimraf": "^4.1.4",
-    "typescript": "^5.1.0"
+    "typescript": "^5.2.2"
   },
   "engines": {
     "node": ">=18"
