@@ -9,13 +9,10 @@ const baseConfig =
       undefined;
 
 /** @type {import('@remix-run/dev').AppConfig} */
-module.exports = {
+export default {
   ...baseConfig,
   ignoredRouteFiles: ["**/.*"],
-  // See https://remix.run/docs/en/main/file-conventions/route-files-v2
-  future: {
-    v2_routeConvention: true,
-  }
+
   // add your own custom config here if you want to.
   //
   // See https://remix.run/docs/en/v1/file-conventions/remix-config
